// web/src/api/models/competitorGroupResource.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * AppStoreCat API
 * Open source app intelligence toolkit API. Provides endpoints for app registration, DNA building, store listing tracking, and analytics.
 * OpenAPI spec version: 1.0.0
 */
import type { AppResource } from './appResource';
import type { CompetitorResource } from './competitorResource';

export interface CompetitorGroupResource {
  parent: AppResource;
  competitors: CompetitorResource[];
}