// web/src/api/models/listAllCompetitorsPlatform.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * AppStoreCat API
 * Open source app intelligence toolkit API. Provides endpoints for app registration, DNA building, store listing tracking, and analytics.
 * OpenAPI spec version: 1.0.0
 */

export type ListAllCompetitorsPlatform = typeof ListAllCompetitorsPlatform[keyof typeof ListAllCompetitorsPlatform];


export const ListAllCompetitorsPlatform = {
  ios: 'ios',
  android: 'android',
} as const;